bcmdhd: replace 0-length arrays with flexible arrays
Linux must have all the "fake" flexible array members replaced with
actual C99 flexible array members so that the programmer's intent can
actually be represented in an unambiguous way (see
-fstrict-flex-arrays). Replace the 0-length arrays with flexible
arrays - it requires no special handling, simply removing the "0" in the
array declaration is sufficient. Everything that could go wrong with
this conversion becomes a compile-time failure and we're out of them.

Change-Id: I95d904ff0ac80c5c5fdd860ef424e0142be2c3a2
Bug: 292213903
Signed-off-by: Tudor Ambarus <tudordana@google.com>
Link: https://people.kernel.org/kees/bounded-flexible-arrays-in-c
Signed-off-by: Will McVicker <willmcvicker@google.com>