bcmdhd: comply with the tdls_mgmt() list of parameters for v6.4
Upstream commit c6112046b1a9c130c455ab0bbe74c3f41138693c introduced a
new parameter for int (*tdls_mgmt)(). Comply with the upstream change to
not break the build.

This is an awful way of dealing with code versioning and dependencies,
but stick with it to ease future merges from older kernels.

Fixes: c6112046b1a9 ("wifi: cfg80211: make TDLS management link-aware")
Change-Id: Icdce9a2546a8fde5a2f48609cca261563c626919
Bug: 294386188
Signed-off-by: Tudor Ambarus <tudordana@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>